test: reorganise tests/ into unit/ and integration/ subdirectories
Move all test files from tests/*.rs into tests/unit/ (pure in-memory,
no external services) and tests/integration/ (needs MinIO/S3/Postgres).
Top-level entry files tests/unit.rs and tests/integration.rs use
#[path] attributes to satisfy Cargo's integration-test discovery rules.
