+++ crates/edgesentry-rs/tests/integration.rs
@@ -1,0 +1,3 @@
+#[cfg(feature = "s3")]
+#[path = "integration/s3_integration.rs"]
+mod s3_integration;
