+++ crates/edgesentry-rs/tests/unit.rs
@@ -1,0 +1,10 @@
+#[path = "unit/core_tests.rs"]
+mod core_tests;
+#[path = "unit/network_policy_tests.rs"]
+mod network_policy_tests;
+#[path = "unit/signing_tests.rs"]
+mod signing_tests;
+#[path = "unit/storage_tests.rs"]
+mod storage_tests;
+#[path = "unit/verify_tests.rs"]
+mod verify_tests;
