+++ crates/edgesentry-rs/tests/unit/core_tests.rs
