+++ crates/edgesentry-rs/tests/unit/signing_tests.rs
