+++ crates/edgesentry-rs/tests/unit/storage_tests.rs
