+++ crates/edgesentry-rs/tests/unit/verify_tests.rs
